[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.c
@@ -2,10 +2,10 @@
 
 int main(void)
 {
-    int A, B;
-    int result;
-    scanf("%d %d", &A, &B);
+    long long A, B;
+    long long result;
+    scanf("%lld %lld", &A, &B);
     result = (A+B)*(A-B);
-    printf("%d", result);
+    printf("%lld", result);
     return 0;
 }
